Add basic CI jobs (build, test, format)
We're adding simple CI workflows for checking of the formatting of the code in
the repo (`format.yml`) and for building the application and running the tests
(`dashboard.yml`). We've configured them to be run on every PR change, after PR
merge to `main` and when manually requested. We will also run the
`dashboard.yml` workflow periodically (every night) on a `main` branch to make
sure the tests are executed even if there are no new commits there.