Add a CI job deploying Bitcoin on Base on testnets
As Wormhole does not support yet the Base Sepolia testnet we want to have a way
of deploying the project on Goerli / Base Goerli when needed. The Goerli testnet
was supposed to get deprecated with the end of 2023, but it looks it is still in
pretty good shape and has many active validators (see
https://goerli.beaconcha.in/). Also, according to https://goerli.basescan.org/,
'Base Goerli will be gradually deprecated starting mid-January and fully shut
down by 9th Feb 2024.'. We hope that by that time Wormhole will start to support
Base Sepolia and we'll be able to switch our config to run fully on Sepolia
testnet. Until then all manual runs of the workflow will result with deploying
using Goerli contracts and all pushes to main will result with deploys using
Sepolia contracts (we may need to some mocking and accept that not everything
will be working there).