Add a workflow deploying BoB on mainnet
We're adding a workflow that will build Bitcoin On Base dashboard and deploy it
to the `bitcoinonbase.org` bucket. The workflow will be triggered each time a
new release is published, but the `dashboard-build-deploy` job won't run
straight away - it will need to be first manually approved due to the
environment protection rules.
As part of this PR we're also changing naming of the workflow building and
deploying testnet BoB dashboard.