Don't load Google Project ID using `load-env-variables` action
The value of `project` key in the `Deploy to GCP` step is the same for both
Sepolia and Goerli environments. We don't need to load it using
`keep-network/ci/actions/load-env-variables` action. Besides, we had some
misconfiguration in the previous config causing the action to not retrieve the
expected data.

The syntax we used in the workflow was interpreted not the way we wanted.
Changing to a different notation.