Use `thesis/gcp-storage-bucket-action` from `alpine-version-413.0.0` branch
We change the version of `thesis/gcp-storage-bucket-action` code used to publish
the dapp image to the GCP. The version that we change to builds the action using
the `413.0.0-alpine` version of the `cloud-sdk`. The `413.0.0-alpine` version
was the version published before the next version introduced a problem with the
execution of the `rsync` command for requesters who do not have the
`storage.buckets.get` permission. According to docs later releases fixed the
issue, but it's still not working for us if we use
`thesis/gcp-storage-bucket-action@v3.1.0` action which pulls the latest
`cloud-sdk` releases. We're getting loads of errors like this:
```
AccessDeniedException: 403 ci-upload-dapp@keep-test-f3e0.iam.gserviceaccount.com
does not have storage.objects.create access to the Google Cloud Storage object.
Permission 'storage.objects.create' denied on resource (or it may not exist).
```
So as a quick workaround, we're using the version of the
`thesis/gcp-storage-bucket-action` action which hardcodes the `cloud-sdk`
version to `413.0.0-alpine`, which we haven't had problems with in the past.

More info:
https://cloud.google.com/storage/docs/release-notes#January_19_2023
GoogleCloudPlatform/gsutil#1663
thesis/gcp-storage-bucket-action@ee8b6a9