Revert commit changing version of the GCP storage action used
In the commit 6787e6199ed238736336ba64f01e95bc851fcb2d we've changed the config
of the workflow, thinking this will help with the following error we were
getting:

```
AccessDeniedException: 403 ci-upload-dapp@keep-test-f3e0.iam.gserviceaccount.com
does not have storage.objects.create access to the Google Cloud Storage object.
Permission 'storage.objects.create' denied on resource (or it may not exist).
```

But the cause of the problem lied in a different place - what we actually had to
do was setting Storage Admin role for
`ci-upload-dapp@keep-test-f3e0.iam.gserviceaccount.com` principal in the config
of the `bob.test.threshold.network` bucket.

We're reverting code to the original version.