Temporarily hardcode Chain ID to ID of Goerli testnet
Tempoparily we're using configuration where values of the `SEPOLIA...` secrets
are set to values corresponding to Goerli environment. So by choosing `sepolia`
we deploy to Goerli. For now we're hardcoding the `CHAIN_ID` to ID of Goerli,
as we don't support Sepolia yet.
In the future we can hardcode this to Sepolia's ID or make the value dependable
on the trigger type and `environment input value.