Fix `useFetchDepositDetails` hook reactivity
There is an issue with data returned by this hook. It is not reactive,
once read it's not updated as the process proceeds. This commit resolves
this issue. Moreover requested and finalized transaction hashes are now
gathered basing on deposit key, updated by minting events subscription
and combined with deposit details in the hook mentioned above.

There is an issue with confirmation numbers which should be updated with
polling store effect. From the point of this hook it should be read from
the store slice and watched by `useEffect` hook so `depositData` can could
be updated.