Improve cleanup function in DepositDetails
The current solution might lead to unexpected behaviors, because we reset the
deposit details step based on a lot of dependencies when we should clean it
after `depositKey` is changed.

This is a quick summary of how clean-up function behaves based on dependency
array:
1. No Dependency Array: Cleanup runs after every render, before the next effect
or on unmount, ensuring effects are properly cleaned up in a constantly updating
component.

2. Empty Dependency Array (`[]`): Cleanup runs once, when the component
unmounts, mirroring a setup that only needs to run and clean up a single time
(like componentDidMount and componentWillUnmount).

3. With Dependencies (`[deps]`): Cleanup runs when any dependency changes, prior
to the next effect to handle updated dependencies, and on unmount, allowing for
targeted effect management and resource cleanup based on specific state or prop
changes.

This means that in our current case, the cleanup function will fire again when
(for example) the new confirmation for the bitcoin transaction arrives, thus set
the `depositDetailsStep` to `bitcoin-confirmations` . It will work anyway,
because we have another `useEffect` below with `confirmations` in dependency
array that will set the proper step, but I think I came up with better solution
that would be less bug-prone. I've separated the cleanup function to another
useEffect with only `depositKey` as a dependency. I've also set it to
`undefined` instead of `bitcoin-transactions`.

```
useEffect(() => {
    return () => {
      updateState("depositDetailsStep", undefined)
    }
  }, [depositKey])
```